pocket (chore): refactor common code
